Refactor `invalidate_assessment_status_cache` to better home for reuse
