+++ portal/views/assessment_engine.py
@@ -12,7 +12,9 @@ from ..database import db
 from ..date_tools import FHIR_datetime
 from ..dogpile import dogpile_cache
 from ..extensions import oauth
-from ..models.assessment_status import AssessmentStatus, overall_assessment_status
+from ..models.assessment_status import AssessmentStatus
+from ..models.assessment_status import invalidate_assessment_status_cache
+from ..models.assessment_status import overall_assessment_status
 from ..models.auth import validate_origin
 from ..models.fhir import QuestionnaireResponse, EC, aggregate_responses, generate_qnr_csv
 from ..models.intervention import INTERVENTION
@@ -1307,14 +1309,6 @@ def assessment_add(patient_id):
                     context='assessment')
     response.update({'message': 'questionnaire response saved successfully'})
 
-    def invalidate_assessment_status_cache(user):
-        """Invalidate the assessment status cache values for this user"""
-        dogpile_cache.invalidate(
-            overall_assessment_status, user.id)
-        for consent in user.all_consents:
-            dogpile_cache.invalidate(
-                overall_assessment_status, user.id, consent.id)
-
     invalidate_assessment_status_cache(patient)
     return jsonify(response)
 
@@ -1322,14 +1316,6 @@ def assessment_add(patient_id):
 @assessment_engine_api.route('/invalidate/<int:user_id>')
 @oauth.require_oauth()
 def invalidate(user_id):
-    def invalidate_assessment_status_cache(user):
-        """Invalidate the assessment status cache values for this user"""
-        dogpile_cache.invalidate(
-            overall_assessment_status, user.id)
-        for consent in user.all_consents:
-            dogpile_cache.invalidate(
-                overall_assessment_status, user.id, consent.id)
-
     user = get_user(user_id)
     if not user:
         abort(404)
